+++ accounts/models.py
@@ -1,25 +1,15 @@
-from django.contrib.auth.models import AbstractBaseUser, BaseUserManager
 from django.db import models
-
-class AccountManager(BaseUserManager):
-    def create_user(self, user_id, password=None, **extra_fields):
-        if not user_id:
-            raise ValueError('The User ID field must be set')
-        user = self.model(user_id=user_id, **extra_fields)
-        user.set_password(password)
-        user.save(using=self._db)
-        return user
-
-    def create_superuser(self, user_id, password=None, **extra_fields):
-        extra_fields.setdefault('is_staff', True)
-        extra_fields.setdefault('is_superuser', True)
-
-        return self.create_user(user_id, password, **extra_fields)
+from django.contrib.auth.models import AbstractBaseUser
+from .managers import AccountManager
+import datetime
 
 class Account(AbstractBaseUser):
     user_id = models.CharField(max_length=255, unique=True)
     name = models.CharField(max_length=255)
     password = models.CharField(max_length=255)
+    created_at = models.DateTimeField(default=datetime.datetime.now)
+    is_staff = models.BooleanField(default=False)
+    is_active = models.BooleanField(default=True)
 
     objects = AccountManager()
