+++ products/ali.py
@@ -1,4 +1,3 @@
-import iop
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from .models import Product
